+++ src/evaluation.cpp
@@ -423,7 +423,8 @@ static int EvaluatePawns(const CPosition *p,
             }
         }
 
-        if (sqCoord.m_nFile < (CSCoord::LEVEL_WIDTH[sqCoord.m_nLevel] - 1) &&
+        if (sqCoord.m_nFile <
+                (static_cast<int>(CSCoord::LEVEL_WIDTH[static_cast<unsigned int>(sqCoord.m_nLevel)]) - 1) &&
             p->m_rgMask[White][Pawn].TstBit(sq + 1)) {
             score += PawnDuo;
         }
@@ -481,7 +482,8 @@ static int EvaluatePawns(const CPosition *p,
             }
         }
 
-        if (sqCoord.m_nFile < (CSCoord::LEVEL_WIDTH[sqCoord.m_nLevel] - 1) &&
+        if (sqCoord.m_nFile <
+                (static_cast<int>(CSCoord::LEVEL_WIDTH[static_cast<unsigned int>(sqCoord.m_nLevel)]) - 1) &&
             p->m_rgMask[Black][Pawn].TstBit(sq + 1)) {
             score -= PawnDuo;
         }
@@ -516,7 +518,8 @@ static int EvaluatePawns(const CPosition *p,
 
     if (tmp_w != tmp_b) {
         tmp_w = tmp_b = 0;
-        for (file = CSCoord::MAX_LEVEL_WIDTH / 2; file < CSCoord::MAX_LEVEL_WIDTH;
+        for (file = static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2);
+             file < static_cast<int>(CSCoord::MAX_LEVEL_WIDTH);
              file++) {
             if (p->m_rgMask[White][Pawn] & FileMask[file])
                 tmp_w++;
@@ -819,7 +822,7 @@ static int EvaluatePassedPawns(const CPosition *p, int wphase, int bphase,
         }
 
         /* Check if 'distant' passed pawn */
-        if (file < (CSCoord::MAX_LEVEL_WIDTH / 2) && !(allpawns & LeftOf[file]) &&
+        if (file < static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) && !(allpawns & LeftOf[file]) &&
             (allpawns & RightOf[file]) &&
             !(p->m_rgMask[Black][Pawn] & LeftOf[file + 2])) {
 #ifdef DEBUG
@@ -831,7 +834,7 @@ static int EvaluatePassedPawns(const CPosition *p, int wphase, int bphase,
             wdistant = MAX(wdistant, rank);
         }
 
-        if (file > ((CSCoord::MAX_LEVEL_WIDTH / 2) - 1) &&
+        if (file > (static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) - 1) &&
             !(allpawns & RightOf[file]) &&
             (allpawns & LeftOf[file]) &&
             !(p->m_rgMask[Black][Pawn] & RightOf[file - 2])) {
@@ -948,7 +951,7 @@ static int EvaluatePassedPawns(const CPosition *p, int wphase, int bphase,
         }
 
         /* Check if 'distant' passed pawn */
-        if (file < (CSCoord::MAX_LEVEL_WIDTH / 2) && !(allpawns & LeftOf[file]) &&
+        if (file < static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) && !(allpawns & LeftOf[file]) &&
             (allpawns & RightOf[file]) &&
             !(p->m_rgMask[White][Pawn] & LeftOf[file + 2])) {
 #ifdef DEBUG
@@ -960,7 +963,7 @@ static int EvaluatePassedPawns(const CPosition *p, int wphase, int bphase,
             bdistant = MAX(bdistant, rank);
         }
 
-        if (file > ((CSCoord::MAX_LEVEL_WIDTH / 2) - 1) &&
+        if (file > (static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) - 1) &&
             !(allpawns & RightOf[file]) &&
             (allpawns & LeftOf[file]) &&
             !(p->m_rgMask[White][Pawn] & RightOf[file - 2])) {
@@ -999,7 +1002,9 @@ static int EvaluatePassedPawns(const CPosition *p, int wphase, int bphase,
 
             while (wrunner) {
                 CSCoord sqCoord = (wrunner).FindSetBitCoord();
-                int dist = (CSCoord::LEVEL_WIDTH[sqCoord.m_nLevel] - 1) - sqCoord.m_nRank;
+                int dist = static_cast<int>(
+                               CSCoord::LEVEL_WIDTH[static_cast<unsigned int>(sqCoord.m_nLevel)]) -
+                           1 - sqCoord.m_nRank;
                 wrunner.ClearLowestBit();
 
                 if (dist < wdist) {
@@ -1797,11 +1802,11 @@ void InitEvaluation(const CPosition *p) {
     int bkfile = p->m_rgKingSq[Black].m_nFile;
     int pawnstorm = 0;
 
-    if (wkfile < ((CSCoord::MAX_LEVEL_WIDTH / 2) - 1) &&
-        bkfile > (CSCoord::MAX_LEVEL_WIDTH / 2)) {
+    if (wkfile < (static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) - 1) &&
+        bkfile > static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2)) {
         pawnstorm = 1;
-    } else if (wkfile > (CSCoord::MAX_LEVEL_WIDTH / 2) &&
-               bkfile < ((CSCoord::MAX_LEVEL_WIDTH / 2) - 1)) {
+    } else if (wkfile > static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) &&
+               bkfile < (static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) - 1)) {
         pawnstorm = 2;
     }
 
@@ -1817,10 +1822,10 @@ void InitEvaluation(const CPosition *p) {
         int wfile = sqCoord.m_nFile;
         int bfile = sqCoord.m_nFile;
 
-        if (p->m_rgKingSq[White].m_nFile < (CSCoord::MAX_LEVEL_WIDTH / 2))
-            wfile = (CSCoord::MAX_LEVEL_WIDTH - 1) - wfile;
-        if (p->m_rgKingSq[Black].m_nFile < (CSCoord::MAX_LEVEL_WIDTH / 2))
-            bfile = (CSCoord::MAX_LEVEL_WIDTH - 1) - bfile;
+        if (p->m_rgKingSq[White].m_nFile < static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2))
+            wfile = static_cast<int>(CSCoord::MAX_LEVEL_WIDTH - 1) - wfile;
+        if (p->m_rgKingSq[Black].m_nFile < static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2))
+            bfile = static_cast<int>(CSCoord::MAX_LEVEL_WIDTH - 1) - bfile;
 
         if (p->m_rgnNonPawn[Black] < eg_threshold) {
             WPawnPos[sq] = (int16_t)(PawnAdvanceEndgame[wfile] * wrank);
@@ -1828,10 +1833,10 @@ void InitEvaluation(const CPosition *p) {
             WPawnPos[sq] = (int16_t)(PawnAdvanceOpening[wfile] * wrank);
         } else {
             WPawnPos[sq] = (int16_t)(PawnAdvanceMiddlegame[wfile] * wrank);
-            if (pawnstorm == 1 && wfile > (CSCoord::MAX_LEVEL_WIDTH / 2)) {
+            if (pawnstorm == 1 && wfile > static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2)) {
                 WPawnPos[sq] += PawnStorm * wrank;
             } else if (pawnstorm == 2 &&
-                       wfile < ((CSCoord::MAX_LEVEL_WIDTH / 2) - 1)) {
+                       wfile < (static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) - 1)) {
                 WPawnPos[sq] += PawnStorm * wrank;
             }
         }
@@ -1843,9 +1848,10 @@ void InitEvaluation(const CPosition *p) {
         } else {
             BPawnPos[sq] = (int16_t)(PawnAdvanceMiddlegame[bfile] * brank);
             if (pawnstorm == 1 &&
-                bfile < ((CSCoord::MAX_LEVEL_WIDTH / 2) - 1)) {
+                bfile < (static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2) - 1)) {
                 BPawnPos[sq] += PawnStorm * brank;
-            } else if (pawnstorm == 2 && bfile > (CSCoord::MAX_LEVEL_WIDTH / 2)) {
+            } else if (pawnstorm == 2 &&
+                       bfile > static_cast<int>(CSCoord::MAX_LEVEL_WIDTH / 2)) {
                 BPawnPos[sq] += PawnStorm * brank;
             }
         }
@@ -1901,7 +1907,7 @@ static bool is_edge(CSCoord coord) {
 }
 
 static void create_mirrored_piece_square_table(int16_t *src, int16_t *dest) {
-    for (int src_idx = 0; src_idx < CSCoord::SIZE; src_idx++) {
+    for (int src_idx = 0; src_idx < static_cast<int>(CSCoord::SIZE); src_idx++) {
         const CSCoord source(src_idx);
         const CSCoord mirrored(0, 7 - source.m_nFile, source.m_nRank);
         dest[static_cast<int>(mirrored)] = src[static_cast<int>(source)];
